Abstract CsvInputFormat
